refactor(admin): extract VisualTimeEntry component
